+++ packages/create-zerotal/templates/react/bootstrap/providers.ts
@@ -1,5 +1,6 @@
 import { LogProvider } from "zerotal/logger";
 import { DatabaseProvider } from "zerotal/orm";
+import { StorageProvider } from "zerotal/storage";
 import { SessionProvider } from "zerotal/session";
 import { AuthProvider } from "zerotal/auth";
 import { InertiaProvider } from "@zerotal/inertia";
@@ -22,6 +23,7 @@ import { DevtoolsProvider } from "@zerotal/devtools";
 const providers = [
   LogProvider,
   DatabaseProvider,
+  StorageProvider,
   SessionProvider,
   AuthProvider,
   InertiaProvider,